backend/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.23.1
sqlmodel==0.0.8
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.24.1
PyJWT==2.8.0
alembic==1.11.1
